+++ src/test/java/TestStats.java
@@ -4,12 +4,12 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TestStats {
 
-    private static final Stats stats = new Stats(new byte[] {2, 4, 6, 2, 1, 17});
-    private static final Stats mods = new Stats(new byte[] {1, 2, 12, 3, 10, 2});
+    private static final Stats stats = new Stats(new int[] {2, 4, 6, 2, 1, 17});
+    private static final Stats mods = new Stats(new int[] {1, 2, 12, 3, 10, 2});
 
     @Test
     void testAddModifiers() {
-        Stats expected = new Stats(new byte[] {3, 6, 18, 5, 11, 19});
+        Stats expected = new Stats(new int[] {3, 6, 18, 5, 11, 19});
         stats.addModifiers(mods);
         assertTrue(equals(stats, expected));
     }
